added new options
Added option to give commands that should be run before the script is
start. this provides a workaround for the Bash shell, which is not
echoing commands for some reason. Running `reset` at the beginning of
the script or with --command reset fixes this.

Added keyboard command 'x' to exit script. Previously Ctl-c was
required.

Added support for reading options from config files. This is supported
by the boost program options library. `gsc` now looks for .gscrc in the
current directory and the home directory and reads them if they exit.

$HOME/.gscrc options will be overridden by ./.gscrc options, which will
be overriden by command line options.