added support disabling implicit return characters
After each script line is loaded, gsc sends a return character (\r) to
the shell so that the command line will be executed. However, if you are
demoing an application like vim, these return characters cause problems.

Added a new implicit_return_mode that can be enabled/disabled with
implicit_return on/off commands. By default, this mode is enabled.