feat: added support for setting window size.
When session starts up, the window size of the psuedoterminal is set to
the size of the current terminal.

Added 'r' command to command mode that syncs the psuedoterminal size to
the current terminal.

So, when gsc starts up, the size of the psuedoterminal matches the
current terminal. If the current terminal size changes, the user can
re-sync the psuedoterminal by pressing 'r' in command mode.