feat: added support for sending multi-character keys together
Added ability to detect and sent multiple character to shell session at
once. For example, pressing the up-arrow key sends OA to the terminal.
It is possible to type these characters into a script using Ctrl-v in
vi, but they appear as multiple characters. So when gsc goes through a
script line, it does not read them together.

gsc now checks to see if a recognized key grouping is coming up and
sends them together if so.