feat: added support for setting context variables
Added ability to specify "context variables", which can be used to
render script files from a template. For example, the user could write

mkdir %semester%-%year%

in their script file and specify

--context-variable="'semester'='Spring'" --context-variable="'year'='2019'"

on the command line to have

mkdir Spring-2019

show up during the session.

This is useful for running commands that depend on the date. For
example, while teaching a class.