feat: added a command parser
Commands can now be embedded in the session scripts. Currently only two
commands are supported, COMMENT and RUN. This allows a script to do
something like

ls
mkdir tmp-dir

for example to make sure a directory or file that your demo will be
creating is removed before you try to create it.

This also changed the behavior of comments. Now comments must be marked

The reason for this is that there are times when you want the session to
load a line that starts with a #. For example, if you are trying to add
a shebang to a file you are editing.