feat: added auto mode
Added auto as a full blown mode with keybindings so that the user can
turn auto on and off during a sesstion. Also added a flag for "full" or
"semi" auto mode. In "full" auto mode, gsc just sends chars to the
shell. In "semi" auto mode, gsc waits for the user to press enter after
a line has been loaded.