fix: fixed bug with multi-char keys causing terminal to stop outputing
this fix is a hack. the problem is that if a key-press results in
multiple characters, each character is processed individually. many
(all?) of these begin with an escape, so this would put gsc into command
mode and the following characters were interpret as if the user pressed
additional key.

so, right now we are just ignoring multiple character key pressed in
insert and command mode. there should be a better implementation in the
future.

on a side note, this may be the cause of the vim problem.