fix: fixes #13 CPU usage
`gsc` was taking up a lot of CPU, and the reason was that, while we were
using select and poll, we were doing so with async calls that would
return immediately. So we basically just had hot loops.

Adding a 100 millisecond timeout fixed the issue, and actually doesn't
have a negative impact on the application because select and poll will
always return as soon as some input is available.

Basically, I just didn't know how to use them correctly...